다빈/1.메인페이지 팝업 js수정(main.js->mainpage.js로 이동) 2. 이벤트팝업 이미지 수정 3. 이벤트페이지 추가/0608
