+++ src/main/webapp/resources/js/main.js
@@ -48,7 +48,6 @@ function getCookie(name) {
 
 // 상단배너, 이벤트팝업 닫기
 const mainBanner = document.getElementById("mainpage-top-banner");
-const eventPopup = document.querySelectorAll(".mainpage-event-popup");
 
 
 // 배너 닫기 
@@ -64,32 +63,16 @@ document.getElementById("header-hide-banner").addEventListener("click",()=>{
 });
 
 
-// 팝업 일반 닫기 
-document.getElementById("mainpage-event-popup-close-btn").addEventListener("click",()=>{
-    eventPopup[0].style.display = "none";
-})  
-
-// 팝업 오늘하루 닫기 
-document.getElementById("mainpage-event-popup-nottoday-btn").addEventListener("click",()=>{
-    // 배너 닫기 
-    eventPopup[0].style.display = "none";
-
-    // hideBanner라는 이름 쿠키 설정(유효기간: 1일) 
-    setCookie("hidePopup", "true", 1);
-});
-
-
 
 // 페이지 로드 시 쿠키 확인하여 요소 숨기기 
 let hideBanner = getCookie("hideBanner");
-let hidePopup = getCookie("hidePopup");
+
 if(hideBanner === "true"){
     mainBanner.style.height="0px"
     mainBanner.style.overflow = "hidden";
 }
-if(hidePopup === "true"){
-    eventPopup[0].style.display = "none";
-}
+
+
 
 
 // 상단배너 end -------------------------------------------------------
